refactor(world): Thorn hazard publishes correct damage batch

+++ Assets/_Scripts/1. Foundation/Events/ProjectileFireArgs.cs
@@ -13,6 +13,6 @@ namespace Foundation
         public int   HomingCount    = 0;
 
         public AudioEventSO[] PierceSounds;
-        public float[] PierceHitStops;
+        public int[] PierceHitStops;
     }
 }
